+++ ricaun.Revit.UI/ConsoleAttribute.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ricaun.Revit.UI
+{
+    /// <summary>
+    /// Console Attribute for ricaun.Console
+    /// </summary>
+    public class ConsoleAttribute : Attribute
+    {
+    }
+}