+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.5] / 2022-01-13
+- Clear Code
 - Change to Now.Tick
+- Add ConsoleAttribute
 - Add RibbonItem Set
 - Add RibbonDescriptionExtension